fix(HA-2):remove auth-form component and add form to auth component
