+++ src/app/auth/auth.component.scss
@@ -1,0 +1,47 @@
+.form-wrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  padding: 1rem;
+}
+
+.form-container {
+  width: 100%;
+  max-width: 400px;
+  min-height: 600px;
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 16px;
+  border-radius: 5px;
+  box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
+}
+
+mat-card {
+  width: 100%;
+  background-color: red;
+  text-align: center;
+  color: white;
+}
+
+.signup-container {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+}
+
+p {
+  margin: 0;
+}
+
+mat-form-field {
+  width: 100%;
+}
+
+button,
+mat-checkbox {
+  align-self: flex-start;
+}
