+++ src/app/auth/pages/login/login.component.html
@@ -1,9 +1,8 @@
-<app-form
-  class="form-wrapper"
+<app-auth
   [serviceError]="serviceError"
   title="Sign in For Work"
   buttonText="Sign in"
   linkText="Don't have an account?"
   [isLoading]="isLoading"
   (submitForm)="onSignIn($event)"
-></app-form>
+></app-auth>
