--- src/app/auth/pages/login/login.component.scss
@@ -1,12 +1,0 @@
-:host {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  padding: 1rem;
-}
-
-.form-wrapper {
-  width: 100%;
-  max-width: 400px;
-}
