+++ src/app/auth/pages/registration/registration.component.html
@@ -1,5 +1,4 @@
-<app-form
-  class="form-wrapper"
+<app-auth
   title="Registration for access"
   [serviceError]="serviceError"
   buttonText="Sign up"
@@ -7,4 +6,4 @@
   [isSignUp]="true"
   [isLoading]="isLoading"
   (submitForm)="onSignUp($event)"
-></app-form>
+></app-auth>
