--- src/app/auth/shared/auth-form/auth-form.component.html
@@ -1,64 +1,0 @@
-<form [formGroup]="form" (ngSubmit)="onSubmit()" class="form-container">
-  <h1>Home bookkeeping</h1>
-  <h3>{{ title }}</h3>
-
-  @if (serviceError) {
-  <mat-card appearance="outlined">
-    <mat-card-content>{{ serviceError }}</mat-card-content>
-  </mat-card>
-  }
-
-  <mat-form-field>
-    <mat-label>Email</mat-label>
-    <input matInput formControlName="email" />
-    @if (getErrorMessage('email')) {
-    <mat-error>{{ getErrorMessage("email") }}</mat-error>
-    }
-  </mat-form-field>
-
-  <mat-form-field>
-    <mat-label>Password</mat-label>
-    <input matInput type="password" formControlName="password" />
-    @if (getErrorMessage('password')) {
-    <mat-error>{{ getErrorMessage("password") }}</mat-error>
-    }
-  </mat-form-field>
-
-  @if (isSignUp) {
-  <mat-form-field>
-    <mat-label>Name</mat-label>
-    <input matInput type="text" formControlName="name" />
-    @if (getErrorMessage('name')) {
-    <mat-error>{{ getErrorMessage("name") }}</mat-error>
-    }
-  </mat-form-field>
-
-  <mat-checkbox formControlName="agree"> I agree with the rules </mat-checkbox>
-  }
-
-  <button
-    mat-raised-button
-    color="primary"
-    type="submit"
-    [disabled]="form.invalid || isLoading"
-  >
-    {{ buttonText }}
-  </button>
-  <div class="signup-container">
-    <p>{{ linkText }}</p>
-
-    <a
-      mat-button
-      color="primary"
-      [routerLink]="isSignUp ? '/auth/login' : '/auth/signup'"
-    >
-      {{ isSignUp ? "Sign in" : "Sign up" }}
-    </a>
-  </div>
-  @if (isLoading) {
-  <mat-progress-spinner
-    mode="indeterminate"
-    diameter="72"
-  ></mat-progress-spinner>
-  }
-</form>
