--- src/app/auth/shared/auth-form/auth-form.component.scss
@@ -1,37 +1,0 @@
-.form-container {
-  min-height: 600px;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 16px;
-  border-radius: 5px;
-  box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
-}
-
-mat-card {
-  width: 100%;
-  background-color: red;
-  text-align: center;
-  color:white;
-}
-
-.signup-container {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
-
-p {
-  margin: 0;
-}
-
-mat-form-field {
-  width: 100%;
-}
-
-button, mat-checkbox {
-  align-self: flex-start;
-}
-
